separated out logic from tech-tree
